New class organisation for idefix: - Hydro is now a sub-class of the dataBlock. All of the "fluid" variables are now in the Hydro class, while the grid info remains in datablock - Conversely, datablock construction automatically triggers a hydro constructor for the said datablock - Riemann solvers have been integrated as Hydro methods.
This allows one to have in principle several datablocks in the same
process, each with their own Hydro object (think of AMR). It also allows
one to have new classes hosted in datablock for new physics which is not
handled by Hydro (Radiative transfer, particles, etc).

Finally, it's cleaner since one can add arrays private to hydro which
need not be allocated by the datablock, as it used to be.

In other words, it's cleaner.