change of tolerance for AmbipolarCshock3D
Tolerance are expected to be slightly affected in v2.0 because v2.0 does
not store the cell centered- fields in its dump, while v1.0 did.

As a result, the error is computed on the remaining fields, and hence is
slightly larger (because cell-centered fields are an average, so the
error due to those is expected to be half).